FOP-2648: Code cleanup
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1761019 13f79535-47bb-0310-9956-ffa450edef68

+++ fop-core/src/main/java/org/apache/fop/afp/AFPResourceLevelDefaults.java
@@ -47,7 +47,7 @@ public class AFPResourceLevelDefaults {
         if (result == null) {
             throw new IllegalArgumentException("Unknown resource type name: " + resourceTypeName);
         }
-        return result.byteValue();
+        return result;
     }
 
     private Map defaultResourceLevels = new java.util.HashMap();
